requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements-dev.in
#
annoy==1.17.0
    # via -r requirements.txt
anyio==3.4.0
    # via
    #   -r requirements.txt
    #   starlette
asgiref==3.4.1
    # via
    #   -r requirements.txt
    #   uvicorn
bidict==0.21.4
    # via -r requirements.txt
bump2version==1.0.1
    # via bumpversion
bumpversion==0.6.0
    # via -r requirements-dev.in
certifi==2021.10.8
    # via
    #   -r requirements.txt
    #   requests
charset-normalizer==2.0.9
    # via
    #   -r requirements.txt
    #   requests
click==8.0.3
    # via
    #   -r requirements.txt
    #   pip-tools
    #   uvicorn
cycler==0.11.0
    # via
    #   -r requirements.txt
    #   matplotlib
dnspython==2.1.0
    # via
    #   -r requirements.txt
    #   email-validator
email-validator==1.1.3
    # via
    #   -r requirements.txt
    #   fastapi
environs==9.4.0
    # via -r requirements.txt
fastapi[all]==0.70.1
    # via -r requirements.txt
fonttools==4.28.4
    # via
    #   -r requirements.txt
    #   matplotlib
h11==0.12.0
    # via
    #   -r requirements.txt
    #   uvicorn
httptools==0.2.0
    # via
    #   -r requirements.txt
    #   uvicorn
idna==3.3
    # via
    #   -r requirements.txt
    #   anyio
    #   email-validator
    #   requests
itsdangerous==2.0.1
    # via
    #   -r requirements.txt
    #   fastapi
jinja2==3.0.3
    # via
    #   -r requirements.txt
    #   fastapi
kiwisolver==1.3.2
    # via
    #   -r requirements.txt
    #   matplotlib
markupsafe==2.0.1
    # via
    #   -r requirements.txt
    #   jinja2
marshmallow==3.14.1
    # via
    #   -r requirements.txt
    #   environs
matplotlib==3.5.1
    # via -r requirements.txt
mongoengine==0.23.1
    # via -r requirements.txt
numpy==1.21.4
    # via
    #   -r requirements.txt
    #   matplotlib
    #   torchvision
orjson==3.6.5
    # via
    #   -r requirements.txt
    #   fastapi
packaging==21.3
    # via
    #   -r requirements.txt
    #   matplotlib
pep517==0.12.0
    # via pip-tools
pillow==8.4.0
    # via
    #   -r requirements.txt
    #   matplotlib
    #   torchvision
pip-tools==6.4.0
    # via -r requirements-dev.in
pydantic==1.8.2
    # via
    #   -r requirements.txt
    #   fastapi
pymongo==3.12.3
    # via
    #   -r requirements.txt
    #   mongoengine
pyparsing==3.0.6
    # via
    #   -r requirements.txt
    #   matplotlib
    #   packaging
python-dateutil==2.8.2
    # via
    #   -r requirements.txt
    #   matplotlib
python-dotenv==0.19.2
    # via
    #   -r requirements.txt
    #   environs
    #   uvicorn
python-multipart==0.0.5
    # via
    #   -r requirements.txt
    #   fastapi
pyyaml==5.4.1
    # via
    #   -r requirements.txt
    #   fastapi
    #   uvicorn
requests==2.26.0
    # via
    #   -r requirements.txt
    #   fastapi
six==1.16.0
    # via
    #   -r requirements.txt
    #   python-dateutil
    #   python-multipart
sniffio==1.2.0
    # via
    #   -r requirements.txt
    #   anyio
starlette==0.16.0
    # via
    #   -r requirements.txt
    #   fastapi
tomli==2.0.1
    # via pep517
torch==1.10.1
    # via
    #   -r requirements.txt
    #   torchvision
torchvision==0.11.2
    # via -r requirements.txt
typing-extensions==4.0.1
    # via
    #   -r requirements.txt
    #   pydantic
    #   torch
ujson==4.3.0
    # via
    #   -r requirements.txt
    #   fastapi
urllib3==1.26.7
    # via
    #   -r requirements.txt
    #   requests
uvicorn[standard]==0.15.0
    # via
    #   -r requirements.txt
    #   fastapi
uvloop==0.16.0
    # via
    #   -r requirements.txt
    #   uvicorn
watchgod==0.7
    # via
    #   -r requirements.txt
    #   uvicorn
websockets==10.1
    # via
    #   -r requirements.txt
    #   uvicorn
wheel==0.37.1
    # via pip-tools
Loko-extensions
    #Loko 

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
